+++ src/Error.php
@@ -28,12 +28,12 @@ class Error extends PrettyJsonSerializable
      *
      * @todo check member constraints: code and error must be set properly
      */
-     public function __construct($code, $error, $message, $description=null, $uri=null)
-     {
-         $this->code        = $code;
-         $this->error       = $error;
-         $this->message     = $message;
-         $this->description = $description;
-         $this->uri         = $uri;
-     }
+    public function __construct($code, $error, $message, $description=null, $uri=null)
+    {
+        $this->code        = $code;
+        $this->error       = $error;
+        $this->message     = $message;
+        $this->description = $description;
+        $this->uri         = $uri;
+    }
 }